Sowmika | MOBN-657 | Extract save of recurring appointments to seperate controller
